[DROOLS-7478] Clean up StoredObject classes
